+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.7.1] - 2024-10-28
+- Fix:
+  - websocket connecting without access token
+  - clear refresh_token only on 4XX with invalid_grant response
+- Chore:
+  - Update dependencies
+
+## [1.7.0] - 2024-09-10
+- Feat:
+  - notify on refresh token expiry/invalid
+  - support multiple, concurrent tenants
+
 ## [1.6.1] - 2024-06-20
 - Chore:
   - update ws version (high vulnerability)